Hide input field / prevent opening keyboard iOS
On Android-Devices (Chrome Mobile) the keyboard hides after a few seconds if I touch the input field. I think it's not necessary and it will be the best to keep Mindmapp simple. To prevent the keyboard from opening every time I select a color picker in Safari (iOS), I added "read-only".